requirements.txt
crewai
langchain-google-genai
python-dotenv
requests
pydantic
sqlite3
